Consolidating the Framework and LocalTypes projects into a single Core project
