+++ BusinessLogic/BusinessLogic.csproj
@@ -47,13 +47,9 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\FrameWork\FrameWork.csproj">
-      <Project>{89c7a321-b6dc-455a-b99f-594a9570fb20}</Project>
-      <Name>FrameWork</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\LocalTypes\LocalTypes.csproj">
-      <Project>{977477BF-6CCA-4FFE-AA4B-B77F9B7D2C68}</Project>
-      <Name>LocalTypes</Name>
+    <ProjectReference Include="..\MyHome2013.Core\MyHome2013.Core.csproj">
+      <Project>{09ae15c4-9db0-4ca8-83bb-eb3b6dea6a55}</Project>
+      <Name>MyHome2013.Core</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
