+++ DataAccess/DataAccess.csproj
@@ -66,18 +66,14 @@
     <Compile Include="SqlMapper.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\FrameWork\FrameWork.csproj">
-      <Project>{89C7A321-B6DC-455A-B99F-594A9570FB20}</Project>
-      <Name>FrameWork</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\LocalTypes\LocalTypes.csproj">
-      <Project>{977477BF-6CCA-4FFE-AA4B-B77F9B7D2C68}</Project>
-      <Name>LocalTypes</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
     <None Include="app.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\MyHome2013.Core\MyHome2013.Core.csproj">
+      <Project>{09ae15c4-9db0-4ca8-83bb-eb3b6dea6a55}</Project>
+      <Name>MyHome2013.Core</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
