+++ DataAccess/PaymentMethodAccess.cs
@@ -1,5 +1,5 @@
 ﻿using Dapper;
-using LocalTypes;
+using MyHome2013.Core.LocalTypes;
 using System.Collections.Generic;
 using System.Data.Common;
 using System.Linq;
