+++ MyHome2013.Core/LocalTypes/Income.cs
@@ -1,54 +1,54 @@
 ﻿using System;
 
-namespace LocalTypes
+namespace MyHome2013.Core.LocalTypes
 {
-    public class Expense
+    public class Income
     {
          #region Properties
 
         /// <summary>
-        /// The amount of the expense
+        /// The amount of the income
         /// </summary>
         public double Amount { get; set; }
 
         /// <summary>
-        /// The date of the expense
+        /// The date of the income
         /// </summary>
         public DateTime Date { get; set; }
 
         /// <summary>
-        /// Category of the expense
+        /// Category of the income
         /// </summary>
-        public ExpenseCategory Category { get; set; }
+        public IncomeCategory Category { get; set; }
 
         /// <summary>
-        /// How the expense was payed
+        /// How the income was payed
         /// </summary>
         public PaymentMethod Method { get; set; }
 
         /// <summary>
-        /// Additional info about the expense
+        /// Additional info about the income
         /// </summary>
         public string Comment { get; set; }
 
         /// <summary>
-        /// ID number of the expense in the data table
+        /// ID number of the income in the data table
         /// </summary>
         public int ID { get; set; }
 
         #endregion
 
         #region C'Tor
-
-        public Expense()
+        
+        public Income()
         {
         }
 
-        public Expense(double amount, DateTime date, ExpenseCategory expenseCategory,
+        public Income(double amount, DateTime date, IncomeCategory incomeCategory,
             PaymentMethod paymentMethod, string comment, int id = 0)
         {
             this.Amount = amount;
-            this.Category = expenseCategory;
+            this.Category = incomeCategory;
             this.Comment = comment;
             this.Date = date;
             this.ID = id;
@@ -61,14 +61,14 @@ namespace LocalTypes
 
         public override bool Equals(object obj)
         {
-            Expense expenseComparing = (Expense)obj;
+            Income incomeComparing = (Income)obj;
 
-            return ((this.Amount == expenseComparing.Amount) &&
-                    (this.Category.Equals(expenseComparing.Category)) &&
-                    (this.Comment == expenseComparing.Comment) &&
-                    (this.Date == expenseComparing.Date) &&
-                    (this.ID == expenseComparing.ID) &&
-                    (this.Method.Equals(expenseComparing.Method)));
+            return ((this.Amount == incomeComparing.Amount) &&
+                    (this.Category.Equals(incomeComparing.Category)) &&
+                    (this.Comment == incomeComparing.Comment) &&
+                    (this.Date == incomeComparing.Date) &&
+                    (this.ID == incomeComparing.ID) &&
+                    (this.Method.Equals(incomeComparing.Method)));
         }
 
         public override int GetHashCode()
