+++ UI/HelperMethods.cs
@@ -1,4 +1,4 @@
-﻿using LocalTypes;
+﻿using MyHome2013.Core.LocalTypes;
 using System.Collections.Generic;
 using System.Linq;
