+++ UI/InputOutUI.cs
@@ -1,5 +1,5 @@
 ﻿using BusinessLogic;
-using LocalTypes;
+using MyHome2013.Core.LocalTypes;
 using System;
 using System.Windows.Forms;
