+++ UI/MyHome2013.csproj
@@ -262,13 +262,9 @@
       <Project>{306B249A-744F-4C66-9365-07CB3EC58A7A}</Project>
       <Name>DataAccess</Name>
     </ProjectReference>
-    <ProjectReference Include="..\FrameWork\FrameWork.csproj">
-      <Project>{89C7A321-B6DC-455A-B99F-594A9570FB20}</Project>
-      <Name>FrameWork</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\LocalTypes\LocalTypes.csproj">
-      <Project>{977477BF-6CCA-4FFE-AA4B-B77F9B7D2C68}</Project>
-      <Name>LocalTypes</Name>
+    <ProjectReference Include="..\MyHome2013.Core\MyHome2013.Core.csproj">
+      <Project>{09ae15c4-9db0-4ca8-83bb-eb3b6dea6a55}</Project>
+      <Name>MyHome2013.Core</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
